OSD/commands: use osdlevel=3 for osd_show_progression command
The osd_show_progression command ('P' key) created text similar to
what --osdlevel=3 shows, and set that as an OSD message. This message
was static and wasn't updated while visible, even if video position
changed (very noticeable during fast forward, when real OSD position
changes rapidly). Instead of setting a static message, create a new
message type that makes the OSD update code behave as if osd level was
set to 3 for the duration of the message. This also avoids having a
duplicated version of the position text formatting logic.