af_scaletempo: fix crash after channel reconfiguration
af_scaletempo could crash if the audio format was changed with
AF_CONTROL_REINIT while the filter was active. In a test case this
happened when audio channels were reconfigured from 6 channels to 2
and playback speed was set to 2. The problem is that reinitialization
does not properly reset filter state. Thus the amount of buffered
audio can be larger than the new maximum buffer size, i.e. the
invariant bytes_queued <= bytes_queue is violated. As a result, memcpy
is called with a negative size.

Fix this by resetting the buffer length when reconfiguring (set the
bytes_queued variable to 0). Also reset some other state for clarity
and robustness, although these changes aren't strictly needed for
avoiding the actual crash.

This may also get rid of some noise played after reinitialization,
caused by using the old buffer contents which were now in the wrong
audio format.