mixer: fix crash when muted and audio init fails
uninit_player() called both mixer_uninit() and ao_uninit() under
"if (mpctx->ao)". However, if AO or filter chain initialization fails
in reinit_audio_chain(), then mpctx->ao is set but mpctx->mixer->ao is
not. mixer_uninit() crashed in this case. Make mixer_uninit() a no-op
if no AO is set in the mixer. Also make the call to mixer_unit() in
uninit_player() unconditional, as calling it is now always safe.