core: clean up libquvi handling, add media_title property
Clean up handling of libquvi (which is used to resolve stream site
URLs into URLs to the underlying media files that can be played
directly). Before, the stream layer tried to do the URL mapping for
every opened stream. Move the code from open.c to quvi.c, and invoke
it explicitly from mplayer.c for the main playback file only.

This allows easily passing metadata from the quvi context to the
frontend. Expose QUVIPROP_PAGETITLE as new "media_title" property (for
YouTube, this is the video title). If libquvi is not used, this is
equal to the old "filename" property. Add "stream_path" property which
is the real played filename/URL after possible libquvi mapping (in
contrast, the old "path" property is the filename/URL that was
originally given to the player).

Handle a potential reliability issue: check quvi_getprop return
values.

Since there was barely anything left in open.c but the open_stream()
stub, move that to stream.c and delete open.c.