configure: fix --language=all
The configure code determining available manpage languages was missing
changes needed after moving the manpage to RST format. As a result,
trying to enable all languages created an invalid list of language
names, which led to a failure when trying to run "make install". Fix.
Note that currently the manpage is not available in the new format in
languages other than English, so trying to enable all languages
doesn't actually do anything different from the default.