cache: saner automatic cache enable; avoids -playlist crash
Code enabling the cache by default for network streams did that by
modifying the value of the "cache" option. This wasn't sane, as
multiple streams may be created and all share the same options. Change
the code to not modify options but store data in the stream instance
instead.

As a side effect, this commit avoids a crash triggered by commit
5722fa1552 ("core, timeline: cache external ordered chapter files
too"). Some existing code creating streams leaves the option pointer
at NULL. That commit modified the function setting cache size for
network streams to use the option pointer, which led to a crash when
using -playlist with a network URL (would also happen in rare cases
like network --sub). This commit removes the pointer use again,
avoiding the crash; the pointer should be set properly though.