vo_gl3: improve Mesa compatibility of context creation
Don't request CONTEXT_FORWARD_COMPATIBLE when creating a GL3 context
on X11. This improves compatibility with some Mesa drivers.

Setting this bit was originally intended to make the code compatible
with newer (future, possibly not yet existing) OpenGL implementations.
These implementations wouldn't have to support deprecated legacy GL
features, which we know that we don't use, and could return a context
with a higher OpenGL version than requested.

In practice, this didn't work out. CONTEXT_FORWARD_COMPATIBLE requires
the driver to explicitly disable deprecated functionality. If it
doesn't do that, it has to reject context creation with that flag set.
And it turns out there are Mesa drivers which suffer from this caveat.
See [1].

This probably also means that a GL application with a fixed set of
required GL features (such as a subset of GL 3.2 core, like vo_gl3.c)
may need to probe several GL versions if drivers really start dropping
legacy GL.

On Windows, CONTEXT_FORWARD_COMPATIBLE is still set. It is not known
if this is ideal, but fortunately there is no Mesa on Windows.

CONTEXT_CORE_PROFILE is still always set. For requested GL versions
lower than 3.2, this flag is ignored. Should we require higher
versions later, we want it to be set, so don't remove it.

[1] https://bugs.freedesktop.org/show_bug.cgi?id=57241