demux_lavf: drop "mp:" prefix from filename given to lavf
Opening files with the libavformat AVISynth demuxer ("avs"/avisynth.c)
failed, because the filename we pass to avformat_open_input() was
prefixed with "mp:". Normally, the name doesn't matter, because lavf
reads data with the stream interface. The AVISynth demuxer can't use
this, because the Avisynth API (apparently) can't read scripts from
memory, and requires a filename.

The "mp:" prefix used to be required when mplayer's stream layer was
made available as protocol to ffmpeg. This was replaced by setting
custom stream callbacks in de4908b644 (svn commit 25499), but the
prefix wasn't removed. Since this prefix doesn't have any purpose
anymore and prevents AVS playback from functioning, remove it.