subs: fix vf_ass --subdelay, remove sub_step slave command
Commit 5639702953 ("subs: always use sub decoder framework for libass
rendering") broke --subdelay with vf_ass (it had no effect). Fix.

Remove sub_step slave command (was bound to keys 'g'/'y' by default).
This command allowed displaying the previous/next subtitle, but was
fairly useless as it was implemented by adjusting sub delay by an
amount that made the correct subtitle appear. This changed the delay
setting permanently, and there was no easy way to change it back to
the previous value.

Move the sub_delay variable to the option struct.