stream_netstream: Remove netstream support
This allowed to move the input stream layer across the network,
allowing the user to play anything that mplayer could play remotely.
For example, playing a DVD related on a remote server (say, with the
host name "remotehost1") could be done by starting the netstream
server (from TOOLS/netstream.c) on that remote server, and then
running:

    mplayer mpst://remotehost1/dvd://

This would open the DVD on the remote host, and transfer the raw DVD
sector reads over network. It works the same for other protocols, and
all accesses to the stream layer are marshaled over network. It's
comparable to the way the cache layer (--cache) works.

It has questionable use and most likely was barely used at all.
There's lots of potential for breakage, because it doesn't translate
the stream CTRLs to network packets. Just get rid of it.