configure: go back to platform-based EXTERN_PREFIX test
Commit 6ad1f58ab8 (svn r28151) changed EXTERN_PREFIX detection to use
a "general" configure test instead of determining it based on
platform. However, this "general" test depends on assumptions about
object file contents, and actually fails if compiling with gcc -flto
for example. Go back to the simple platform-based test. Leave out the
OS2 (no longer supported) and OpenBSD !ELF (no idea what that is, but
probably not something I'd care about supporting) parts from the old
test.

Remove configure switch --nm, as it was only used for the hacky
EXTERN_PREFIX test.

Simply disable the use of EXTERN_ASM (same as EXTERN_PREFIX under
another name) in wrapper.S; this seems to have been added for OS/2
only, and the file should not get compiled on any platform with a
nonempty prefix now.