subassconvert: do not escape likely ASS override tags
Some Windows players implement rendering of SubRip files by converting
to ASS format, but fail to escape text that could be considered ASS
markup. Unfortunately, this had led to the creation of files which mix
SubRip and ASS markup, relying on rendering being implemented this way
and on the ASS markup falling through to the ASS renderer unescaped.

Add a simple heuristic to detect whether '{' is likely to start an ASS
tag: if the string starts with '{\', and there is a closing '}',
assume it's an ASS tag and pass it through unescaped, otherwise escape
the '{' properly.

Thus tags like "{\a6}" are interpreted as ASS markup, while most
legitimate uses of '{' and '}' should work fine. We assume that nobody
uses {...} for commenting text in SubRip files. (This kind of comment
is popular and legal in ASS files, though.)